There was a problem with related parents
If you selected an related parent, all tiddlers with this tag were listed,
not just those that were on the current "level" of the hierarchy.

For example in the Ford family the breadcrumbs were

* Henry Ford
* Edsel Ford
* V.P.

...the list was:

* Benson Ford
* Elena Ford <- should not be listed

Elena is in a lower level in the hierarchy, but she was listed, because
she has the "V.P." tag (the last in the breadcrumbs).

Now the last opened "direct tag" is stored in the state tiddler, so the
filter of the list elements basically is:

  [[Last opened direct tag]tag[Related]tag[tags]]