Split search and Locator-view tags
If you want to SEARCH in context, then you don't need to NAVIGATE in the
search results, thus it doesn't needed to share the same context.

If you want to search in current Locator view, then press the "circle"
button, it will copy the tags of context to the search state tiddler,
but all of the direct tags will be converted to ancestor tags: without
this, the search would like to find tiddlers with a lot of direct tags.

I've added buttons to narrow the search by direct/ancestor tags to get
more precise results.

Use "tag icon" instead of arrow for direct tags in breadcrumbs and
search results, because it makes cleaner the kind of tags.