Use the same "style" for breadcrumb- and list items
When we selected an ancestor tag from the list, it became a tag-pill in
breadcrumbs.

This operation was not consistent, now it shows the elements in the same way
in breadcrumbs and in the list.

I use buttons to display titles to appear correctly in the search drop-down
list as well as in the sidebar.

https://groups.google.com/d/msg/tiddlywiki/C_eGviHgLPg/CAeo1QuAGwAJ