Move tiddler from current context to subtopic on dropping
For example, if we look at the children of Edsel Ford (we open Edsel
Ford in the Locator sidebar) and drag the title Josephine Clay Ford to
the title William Clay Ford, the tag Edsel Ford will be removed from
Josephine and replaced with the William Clay Ford tag.

This is useful if you want to re-organize Table of Contents by splitting
a larger topic into several smaller topics.

You can drag-n-drop any tiddler title (even from tiddler text) to the
titles in the Locator sidebar to apply add the target tag to it.