Make relationship-settings bordered; highlight its icon when opened
Intertwined with history, it was difficult to distinguish between them.

Previously, its icon was highlighted only when "invert direction" was
turned on because otherwise I couldn't make it stand out. Since then, I
change the icons when it is turned on, so the toggle icon on the
relationship settings is enough to indicate that the settings are open.