Option to use template of relationship field for context items
For example, if tiddlers containing images or videos are context items,
you can specify a template to display the contents of the tiddler, so
you can actually create a filterable playlist.

I do not use the template in history elements because it can easily
happen that history is full of video players. This would make navigation
difficult.