Search in any `locator-view`
In order to not only use standard search for sidebar Locator, but for
any `locator-view`, I moved the search button to `locator-view` from the
search toolbar.

In fact, I only need to copy a single field, the list of
`opened-field-filters`, along with the active filters, so I modified the
copying of contextState.