"Open in sidebar" appends to history instead of special handling
Previously, the "open in sidebar" in the tiddler toolbar set a field,
and if this field had a value, the Locator sidebar appeared differently,
and had to press a button to return to its original context.

Instead, it's simply enough to add a new element to history so you can
easily get back to where you came from.