"Remove field filters" in search did not cleared recursive search state
When recursive search was turned on for a field value, the button that
turned off all filters at once did not clear recursive search. This did
not affect the behaviour because the recursive search without the
intersection and difference state had no effect, but the recursive
search buttons remained active (red colored) in the field filter.